fix documentation building
An earlier patch added --disable-docs, but went too far, making it
impossible to build the docs.

It turns out that Automake seemingly has a bug preventing the
conditional build of an info file.  So, this patch works around the
bug by putting the info_TEXINFOS rule into a new doc/Makefile.am.

Tested by building with and without --disable-docs and looking for the
existence of doc/libffi.info.