Fix .cfi pseudo-ops test.
The existing tests were failing because the GCC was wrapping the test
asm in cfi directives, which gas doesn't like.  So now we test if the
compiler emits them first, and assume that the assembler will accept
them.  If that doesn't work, then we test if the assembler will accept
them assuming the compiler doesn't emit them.